# docs/swagger.yaml
basePath: /
definitions:
  handlers.AddToCartRequest:
    properties:
      productId:
        type: integer
      quantity:
        type: integer
    type: object
...
  handlers.CreateCarouselRequest:
    properties:
      imageUrl:
        type: string
      linkUrl:
        type: string
      sortOrder:
        type: integer
      title:
        type: string
    type: object
...
  handlers.DeleteCartItemRequest:
    properties:
      itemIds:
        items:
          type: integer
        type: array
    type: object
...
  handlers.UpdateCartItemRequest:
    properties:
      itemId:
        type: integer
      quantity:
        type: integer
    type: object
...
  models.Carousel:
    properties:
      id:
        type: integer
      imageUrl:
        type: string
      linkUrl:
        type: string
      sortOrder:
        type: integer
      status:
        description: active, inactive
        type: string
      title:
        type: string
    type: object
  models.Cart:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      totalItems:
        description: Total number of items
        type: integer
      totalPrice:
        description: Sum of all item subtotals
        type: number
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  models.CartItem:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      price:
        description: Unit price
        type: number
      productId:
        description: Product ID
        type: integer
      productName:
        description: Product name
        type: string
      quantity:
        description: Quantity of this item
        type: integer
      subtotal:
        description: Calculated subtotal (price * quantity)
        type: number
      updatedAt:
        type: string
      userId:
        description: User ID for persistence
        type: integer
    type: object
  models.CartResponse:
    properties:
      cart:
        $ref: '#/definitions/models.Cart'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PetShop 宠物商店 API 文档 - 提供宠物管理、购物车、订单管理、用户管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PetShop API
  version: "1.0"
paths:
...
  /api/admin/carousel:
    delete:
      consumes:
      - application/json
      description: 根据ID删除轮播图
      parameters:
      - description: 轮播图ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 轮播图不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除轮播图
      tags:
      - 轮播图管理
    put:
      consumes:
      - application/json
      description: 更新现有轮播图信息
      parameters:
      - description: 轮播图信息
        in: body
        name: carousel
        required: true
        schema:
          $ref: '#/definitions/models.Carousel'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的轮播图
          schema:
            $ref: '#/definitions/models.Carousel'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 轮播图不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新轮播图
      tags:
      - 轮播图管理
  /api/admin/carousels:
    get:
      consumes:
      - application/json
      description: 获取所有轮播图
      produces:
      - application/json
      responses:
        "200":
          description: 轮播图列表
          schema:
            items:
              $ref: '#/definitions/models.Carousel'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取轮播图列表
      tags:
      - 轮播图管理
    post:
      consumes:
      - application/json
      description: 创建新的轮播图
      parameters:
      - description: 轮播图创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCarouselRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功的轮播图
          schema:
            $ref: '#/definitions/models.Carousel'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 创建轮播图
      tags:
      - 轮播图管理
...
  /api/cart:
    delete:
      consumes:
      - application/json
      description: 从购物车中删除指定商品
      parameters:
      - description: 删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除后的购物车
          schema:
            $ref: '#/definitions/models.CartResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 购物车项不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 从购物车删除商品
      tags:
      - 购物车
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的购物车内容
      produces:
      - application/json
      responses:
        "200":
          description: 购物车信息
          schema:
            $ref: '#/definitions/models.CartResponse'
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取购物车
      tags:
      - 购物车
    post:
      consumes:
      - application/json
      description: 将商品添加到当前用户的购物车
      parameters:
      - description: 添加商品请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddToCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 添加成功后的购物车
          schema:
            $ref: '#/definitions/models.CartResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 添加商品到购物车
      tags:
      - 购物车
    put:
      consumes:
      - application/json
      description: 更新购物车中指定商品的数量
      parameters:
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的购物车
          schema:
            $ref: '#/definitions/models.CartResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 购物车项不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新购物车商品数量
      tags:
      - 购物车
  /api/cart/clear:
    delete:
      consumes:
      - application/json
      description: 清空当前用户的购物车中的所有商品
      produces:
      - application/json
      responses:
        "200":
          description: 清空后的购物车
          schema:
            $ref: '#/definitions/models.CartResponse'
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 清空购物车
      tags:
      - 购物车
...
securityDefinitions:
  APITokenAuth:
    description: 使用 API Token 进行认证，用于 Open API 访问
    in: header
    name: X-API-Token
    type: apiKey
  BearerAuth:
    description: 使用 JWT Token 进行认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"